Fix LE Cert not being generated for www.
The appropriate nginx config block for matching the well known directory
to the appropriate folder was missing in the www. config block.

When it's being deployed, it may be needed to delte the old ssl
certificate in `/etc/letsencrypt/live/`, `/etc/letsencrypt/renewal/`
and `/etc/letsencrypt/archive/cuties.social` manually before deploying
again.