Fix nginx config to re-renew ssl-certs
Due to a error in the nginx config, the SSL Certificate was not renewed
correctly for the www. subdomain. Assuming `mastodon_home` to be
`/home/mastodon/` and `mastodon_path` to be `/home/mastodon/live/`, the
generated nginx configuration would look like this:

```nginx
location /.well-known/acme-challenge/ {
  alias /home/mastodon//home/mastodon/live/public/.well-known/acme-challenge/;
}
```

The fatal flaw being the alias to `/home/mastodon//home/mastodon/live`
and the double-reference to /home/mastodon/. But since this was only
relevant for the www. subdomain (introduced in 4c37d5d by myself) was
hard(er) to find. The nginx error log would hint for this error with
this entry in the error log:

```
[...]
2020/12/05 18:46:57 [error] 13356#13356: *59 open() "/home/mastodon//home/mastodon/live/public/.well-known/acme-challenge/a" failed (2: No such file or directory), client: xxxx:xx:xxxx:xxxx:xxx:xxxx:xxxx:xxxx, server: www.cuties.social, request: "GET /.well-known/acme-challenge/a HTTP/1.1", host: "www.cuties.social"
[...]
```

At the beginning I did not notice the difference between the main domain
and the www. subdomain. After seeing the rate-limit in the logs and
switching to another CA (see e398c4e) I only checked if
`/.well-known/acme-challenge/` was working (placing a file in the
corresponding web-root and retrieving it via curl) for the main domain.
But I failed to check for the www. subdomain. So the lesson learned is:
always check for all domains that are relevant for the certificate
instead of just the main domain. This could have saved some 10 to 15
Minutes of debugging and ~10 Minutes writing this "Lessons-Learned"
Block.