Switch certbot to pip install
The Ubuntu 18.04 certbot version is horribly outdated (0.27 vs. 1.16),
and a error linked to an old version[^1] prevented the renew of the SSL
certificate.
This commit removes the old apt based packages and installs certbot from
pip to get a current version. Due to the rust dependency in the
cryptography pip needs to be upgraded to[^2]. Now our certificated was
renewed like a charm :)

[^1]: https://community.letsencrypt.org/t/problem-with-renew-certificates-the-request-message-was-malformed-method-not-allowed/107889/16
[^2]: https://cryptography.io/en/latest/installation/#installation